requirements.txt
joblib==1.2.0
matplotlib==3.7.1
numpy==1.21.6
pandas==2.0.2
Pillow==10.0.0
plotly==5.15.0
seaborn==0.12.2
streamlit==1.23.1
scikit-learn==1.2.0